CXO-1080	- Remove time of use calendar

+++ coko/com.elster.jupiter.calendar/src/main/java/com/elster/jupiter/calendar/impl/gogo/CalendarCommands.java
@@ -73,7 +73,11 @@ public class CalendarCommands {
     public void deleteCalendar(String name) {
         Optional<Calendar> obsolete = this.calendarService.findCalendarByName(name);
         if (obsolete.isPresent()) {
-            this.deleteCalendar(obsolete.get());
+            if (this.calendarService.isCalendarInUse(obsolete.get())) {
+                System.out.println("Calendar with name " + name + " is in use!");
+            } else {
+                this.deleteCalendar(obsolete.get());
+            }
         } else {
             System.out.println("Calendar with name " + name + " not found!");
         }
